# NIFI-10888: When inferring a schema using a Record Reader, buffer up …