@NissimShiman 2022-12-09
This is a really nice enhancement.

Tested with csv files being converted to json files before and after fix
for csv files of 21 and 63 bytes, there was 25 -30% speed up in processing (19/20 seconds to 15 seconds)
for csv files of 549 bytes, I saw a 33%+ speed up (37 to 23 seconds)

This was compiled/run with java openjdk 1.8.0_332
on linux kernel 3.10.0-1160