@mattyb149 2022-12-14
+1 LGTM, thanks for the improvement! Merging to main